backend/requirements.txt
fastapi==0.135.1
google-adk==1.27.1
google-cloud-aiplatform==1.141.0
google-cloud-firestore==2.25.0
google-genai==1.67.0
uvicorn[standard]==0.42.0
